renamed inter_thread_communicator to inter_task_communicator
